Implement nested coercion
