+++ Gemfile
@@ -2,3 +2,5 @@ source 'https://rubygems.org'
 
 # Specify your gem's dependencies in pragma-contract.gemspec
 gemspec
+
+gem 'pry'
