+++ spec/pragma/contract/base_spec.rb
@@ -7,13 +7,17 @@ RSpec.describe Pragma::Contract::Base do
     Class.new(described_class) do
       property :title, type: coercible(:string)
 
+      collection :songs do
+        property :title, type: coercible(:string)
+      end
+
       validation do
         required(:title).filled
       end
     end
   end
 
-  let(:model) { OpenStruct.new }
+  let(:model) { OpenStruct.new(songs: [OpenStruct.new]) }
 
   it 'is instantiated correctly' do
     expect { subject }.not_to raise_error
@@ -27,5 +31,10 @@ RSpec.describe Pragma::Contract::Base do
   it 'validates properties correctly' do
     subject.validate(title: nil)
     expect(subject).not_to be_valid
+  end
+
+  it 'coerces values in nested properties correctly' do
+    subject.validate(title: 1, songs: [{ title: 2 }])
+    expect(subject.songs.first.title).to eq('2')
   end
 end
