+++ spec/spec_helper.rb
@@ -1,6 +1,8 @@
 require 'coveralls'
 Coveralls.wear!
 
+require 'pry'
+
 $LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
 require "pragma/contract"
